explain,coord: move `explain_materialized_view`
Move `explain_materialized_view` to `inner/create_materialized_view.rs`.
